Added FileBase class, LocalFile now extends it
