+++ src/main/java/edu/pdx/cs510/agile/team3/FTP/RemoteFile.java
@@ -1,0 +1,7 @@
+package edu.pdx.cs510.agile.team3.FTP;
+
+/**
+ * Created by henry on 7/22/17.
+ */
+public class RemoteFile {
+}
